[e2e] create Setup class and refactoring of BaseTest #2438
